Made the "managers" (like TeamMngr) singletons and globally available through static calls
Changed database handling to passing around pointers to one single DB object / instance around in the whole application. The object is owned by the Tournament class; as long as the Tournament instance exists, the database handle exists.

Introduced a method for renaming teams.

Added a check for maximum name lengths to Team creation and rename.