Tested changes: Introduced new classes for handling scores (both games scores and match scores). Class functionality is mostly covered by unit tests.
Completely untested changes:
  * Integration of the new score classes with Match and MatchManager
  * State transitions of Match, MatchGroup, Court and Category upon
entering a match result (and thus finishing a match)
  * Getter method in Match to retrieve the assigned court as an Court
instance

I only know that the untested changes compile. I haven't even tested
them manually.

KNOWN BUG:
The check for the state transition of a category from "RUNNING" to
"FINALIZED" is incomplete. Yet to be generated games of yet-to-be-played
rounds are not taken into account.